% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_at}
\alias{as_at}
\title{Get the 'as at date' of a \code{val_data}}
\usage{
as_at(val_data)
}
\arguments{
\item{val_data}{The valuation data.}
}
\value{
A scalar \code{datey}.
}
\description{
Gets the 'as at' date of a valuation data.
}
