% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{assign_column}
\alias{assign_column}
\alias{$<-.exp_data}
\alias{$<-.val_data}
\title{Add or assign a data column}
\usage{
\method{$}{exp_data}(x, name) <- value

\method{$}{val_data}(x, name) <- value
}
\arguments{
\item{x}{The \code{exp_data} or \code{val_data}.}

\item{name}{The \code{name} of the column to assign.}

\item{value}{The column values.}
}
\value{
The updated \code{exp_data} or \code{val_data}.
}
\description{
Add or assign a data column to \code{exp_data} or \code{val_data}.
}
