% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exp_data_info}
\alias{exp_data_info}
\alias{exp_start}
\alias{exp_end}
\title{Get experience period of \code{exp_data}}
\usage{
exp_start(exp_data)

exp_end(exp_data)
}
\arguments{
\item{exp_data}{The mortality experience data.}
}
\value{
A scalar \code{datey}.
}
\description{
\code{exp_start} gets the start of the overall experience period (inclusive).

\code{exp_end} gets the end of the overall experience period (exclusive).
}
