% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{is_exp_and_val_data_type}
\alias{is_exp_and_val_data_type}
\alias{is_exp_data}
\alias{is_val_data}
\title{Is \code{x} an \code{exp_data} or a \code{val_data}?}
\usage{
is_exp_data(x)

is_val_data(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A scalar \code{logical}.
}
\description{
\code{is_exp_data()} checks whether an object is a \code{exp_data}.

\code{is_val_data()} checks whether an object is a \code{val_data}.
}
