% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmu_function.R
\name{is_logmu_function}
\alias{is_logmu_function}
\title{Test for a logmu function object}
\usage{
is_logmu_function(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
A scalar \code{logical}.
}
\description{
Tests whether \code{x} is any \strong{logmu} function object -- a \code{mortality},
\code{variable}, \code{static_variable}, \code{include} or \code{indicator}.
}
