% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmu-package.R
\docType{package}
\name{logmu-package}
\alias{logmu}
\alias{logmu-package}
\title{Actuarial mortality experience analysis and modelling}
\description{
TODO: CHANGE THIS TO A USER SUMMARY OF THE PACKAGE

NOTE THAT FOR CONINTUOUS TIME \strong{logmu} USES DATEY PACKAGE

The \strong{logmu} package provides
high-performance actuarial mortality experience analysis and
model fitting, built on a flexible mortality framework incorporating
time-based covariates and arbitrary proportional hazards models and
with particular support for postcode-based socio-economic mortality
models.

A/E analysis includes confidence intervals and residuals with visualisation.
Model fitting features include covariate clustering,
maximum likelihood fitting and model selection using AIC.

Weighting #' (e.g. amounts vs lives),
probabilistic similarity weighting (e.g. down-weight older data) and
optionally time-based inclusion criteria (e.g. sub-setting experience by age) are supported throughout.

Calculations use SIMD vectorisation and multi-threading for performance.
}
\seealso{
\itemize{
\item TODO
\item vignette("XXX") for a worked introduction.
\item The series of articles starting with \code{vignette("Measures matter")}
for the \strong{logmu} theoretical framework.
}
}
\author{
\strong{Maintainer}: Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})

Authors:
\itemize{
  \item Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})
}

}
\keyword{internal}
