% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_const.R
\name{mortality_const}
\alias{mortality_const}
\title{Create a constant \code{mortality}}
\usage{
mortality_const(q = NULL, mu = NULL, log_mu = NULL, name = NULL)
}
\arguments{
\item{q}{The probability of dying over one years. It is required that
\eqn{0 < q < 1}.}

\item{mu}{The instantaneous annual rate ('force') of mortality. It is required that
\eqn{0 < \mu < +\infty}.}

\item{log_mu}{The natural logarithm of the instantaneous annual rate ('force') of mortality. It is required that
\eqn{-\infty < \log\mu < +\infty}.}

\item{name}{An optional name for this mortality.}
}
\value{
A \code{mortality}.
}
\description{
Creates a constant mortality from \eqn{q}, \eqn{\mu} or \eqn{\log\mu}.

The annual mortality rate must be provided as one
(and only one) of \code{q}, \code{mu} or \code{log_mu} as appropriate.
}
