% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_table.R
\name{mortality_table}
\alias{mortality_table}
\title{Create an \code{mortality} from a 2D age-period matrix of annual mortality rates}
\usage{
mortality_table(x0, t0, q = NULL, mu = NULL, log_mu = NULL, name = NULL)
}
\arguments{
\item{x0, t0}{The youngest age and earliest time respectively.
For \eqn{q} rates, these are the \emph{middle} of the year covered by
the youngest and earliest \eqn{q} rate, i.e. \eqn{\mu}{mu}-timing.}

\item{q}{An age-period matrix of \eqn{q_{xt}}. It is required that
\eqn{0 < q < 1}.}

\item{mu}{An age-period matrix of \eqn{\mu_{xt}}. It is required that
\eqn{0 < \mu < +\infty}.}

\item{log_mu}{An age-period matrix of \eqn{\log\mu_{xt}}. It is required that
\eqn{-\infty < \log\mu < +\infty}.}

\item{name}{An optional name for this mortality.}
}
\value{
A \code{mortality}.
}
\description{
Creates an age-period mortality table from a 2D age-period matrix of
annual mortality rates that can be \eqn{q_{xt}}, \eqn{\mu_{xt}} or
\eqn{\log\mu_{xt}}.

The first age and period are determined by the \code{x0} and \code{t0} parameters
respectively using \eqn{\mu}{mu}-timing. (See below for what this means for
\eqn{q} rates.)

The end age of the table is determined as the start age \code{x0}
plus the number of age rows as years (plus another half year for
\eqn{q} rates -- see below).

The annual mortality rates must be provided as a 2D age-period matrix in one
(and only one) of \code{q}, \code{mu} or \code{log_mu} as appropriate.

It is an implicit assumption that these rates are 'smooth' at the annual
scale, i.e. the second differences of \eqn{\mu} and \eqn{\log\mu} are
'small', e.g.
\eqn{\Delta^2\mu<10\\\%\mu} and
\eqn{\Delta^2\log\mu<10\\\%}.
If you want to allow for realistic, i.e. non-smooth, historical
annual and sub-annual noise then use a \code{variation}.

Important notes for \eqn{q} rates:
\enumerate{
\item All timing is \eqn{\mu}{mu}-timing. This means:
\itemize{
\item The definition of \eqn{q_{xt}}{q_xt} is \emph{centred} on \eqn{(x,t)}, i.e.

\deqn{q_{xt} = 1 - \exp\left(-\int_{t-\frac{1}{2}}^{t+\frac{1}{2}}\mu_{x+\varepsilon,\, t+\varepsilon}\,\mathrm{d}\varepsilon\right)}{q_xt = 1 - exp(-integral from t-1/2 to t+1/2 of mu_(x+e, t+e) de)}
}

This differs from the normal convention whereby \eqn{q_{xt}}{q_xt} relates
to the year from \eqn{(x,t)} to \eqn{(x+1,t+1)}.
\itemize{
\item The \code{x0} and \code{t0} parameters relate to the \emph{middle} of the year
covered by the youngest and earliest \eqn{q} rate.
\item The end age of the resulting mortality table is the start age \code{x0}
plus the number of age rows as years \emph{plus an extra half year}.
}
\item The calculation of \eqn{\mu_{xt}} from \eqn{q_{xt}} includes an
allowance for estimated convexity determined by examining the two
neighbouring \eqn{q} rates
(by cohort for the interior of the \code{annual_rates} and by age at its edges).
This may produce artefacts if the rates are not smooth at an annual scale.
\item A common convention when specifying \eqn{q}-based mortality tables is to
include \eqn{q_\omega=1}, where \eqn{\omega} is the end age of the mortality
table.
\emph{Do not include a \eqn{q=1} age row in the \code{annual_rates} argument.}
(If you are creating the mortality from a base table and a projection then
the \eqn{q_\omega=1} is likely in the base table.)
}
}
