% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{val_from_exp_data}
\alias{val_from_exp_data}
\title{Create a \code{val_data} from an \code{exp_data}}
\usage{
val_from_exp_data(exp_data, as_at = NULL)
}
\arguments{
\item{exp_data}{The \code{exp_data}.}

\item{as_at}{The as at date (a \code{datey}). If omitted, the \emph{end} of the
experience period is used.}
}
\value{
A \code{val_data}.
}
\description{
Creates a \code{val_data} from those individuals  in an \code{exp_data} who
were alive at the specified date.
}
