% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_ops.R
\name{vec_ops}
\alias{vec_ops}
\alias{vec_neg}
\alias{vec_exp}
\alias{vec_expm1}
\alias{vec_log}
\alias{vec_log1p}
\alias{vec_m_from_q}
\alias{vec_add}
\alias{vec_sub}
\alias{vec_mul}
\alias{vec_div}
\alias{vec_pow}
\alias{vec_min}
\alias{vec_max}
\alias{vec_clamp}
\title{Vector ops}
\usage{
vec_neg(x)

vec_exp(x)

vec_expm1(x)

vec_log(x)

vec_log1p(x)

vec_m_from_q(x)

vec_add(x, y)

vec_sub(x, y)

vec_mul(x, y)

vec_div(x, y)

vec_pow(x, y)

vec_min(x, y)

vec_max(x, y)

vec_clamp(x, min, max)
}
\arguments{
\item{x}{A vector of double.}

\item{y}{A vector of double.}

\item{min, max}{Vectors of double for the \code{clamp} operation.}
}
\value{
A vector of \code{double}.
}
\description{
Vector ops
}
